% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accentuate.R
\name{accentuate}
\alias{accentuate}
\title{Magnification of co-occurrence matrix.}
\usage{
accentuate(correlation_matrix_list)
}
\arguments{
\item{correlation_matrix_list}{An N x N co-occurrence matrix.}
}
\value{
An N x N co-occurrence matrix with increased covariate effects.
}
\description{
Takes co-occurrence matrix and multiplies paffects so that they are easier to view, with the absolute maximum covariate being increased to 1.
}
