% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_deviance.R
\name{calc_deviance}
\alias{calc_deviance}
\title{Calculate deviance}
\usage{
calc_deviance(y, p, log = TRUE)
}
\arguments{
\item{y}{Acutal presences/absence of species}

\item{p}{Predicted likelihood of presence or absence of species.}
}
\value{
Deviance value for model.
}
\description{
Calculates deviance of a given model.
}
