% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_association_gradient.R
\name{create_association_gradient}
\alias{create_association_gradient}
\title{Species associations gradient.}
\usage{
create_association_gradient(association_matrix_list, focal_species)
}
\arguments{
\item{association_matrix_list}{A list of N x N co-occurrence matrices, one for each level on the environmental variable that was emasured.}

\item{focal_species}{The single species you want to observe the changes in associations for}
}
\value{
A species x covariate level matrix.
}
\description{
Takes list of association matrices and creates a new matrix showing the change in species associations over a change in environmental variable for one species.
}
