% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lambda.R
\name{create_lambda}
\alias{create_lambda}
\title{Creates lambda matrix.}
\usage{
create_lambda(n_species, n_latent)
}
\arguments{
\item{n_species}{Number of species involved in model.}

\item{n_latent}{Number of latent variables chosen.}
}
\value{
A species x latent variable GRETA matrix with a uniformly distributed diagonal and an upper right triangle of zeroes.
}
\description{
Creates restrained lambda matrix for us in JSAM model.
}
