% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_ints.R
\name{get_param_ints}
\alias{get_param_ints}
\title{Derive parameter confidence intervals}
\usage{
get_param_ints(parameter, draws, species_names, env_names, ssv = FALSE,
  quantile_n = "mean")
}
\arguments{
\item{parameter}{Parameter for which we want to derive credible intervals.}

\item{draws}{Set of mcmc draws.}

\item{species_names}{Names of species to analyse.}

\item{env_names}{Names of environmental variables.}

\item{ssv}{Whether or not we are trying to extract species-specific variables.}

\item{quantile_n}{0.025, "mean" or 0.975. Default of "mean".}
}
\value{
A table of interval values for our parameters.
}
\description{
Function extracts confidence intervals for our parameters.
}
