% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrowing.R
\name{narrowing}
\alias{narrowing}
\title{Subsetting of co-occurrence matrix.}
\usage{
narrowing(correlation_matrix_list, focal_species_list)
}
\arguments{
\item{correlation_matrix_list}{An N x N co-occurrence matrix.}

\item{focal_species_list}{A vector of M characters which correspond to species names.}
}
\value{
An M x M co-occurrence matrix.
}
\description{
Takes co-occurrence matrix and narrows down to only show relationships between species of interest.
}
