% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_change_asc.R
\name{plot_change_asc}
\alias{plot_change_asc}
\title{Plot species association changes.}
\usage{
plot_change_asc(association_gradient_change_matrix, legend_position = c(2, 1))
}
\arguments{
\item{association_gradient_change_matrix}{A species by environmental covariate level matrix.}
}
\value{
A plot showing changes in species associations.
}
\description{
Takes association change matrix and plots it, visualizing the changes in species association between one focal species and others.
}
