% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_cond.R
\name{pred_cond}
\alias{pred_cond}
\title{Conditional prediction of species at one site.}
\usage{
pred_cond(betas_shared, betas_ssv = NULL, site_env, occupancy, cooccurrence,
  focal_species, ssv = FALSE)
}
\arguments{
\item{betas_shared}{Non species-specific enrivonmental parameters.}

\item{betas_ssv}{Species-specific environmental parameters.}

\item{occupancy}{Occupancy vector for all species.}

\item{cooccurrence}{Cooccurrence matrix for the site (detrmined by the temperature at the site, use pred_correlation to derive this).}

\item{focal_species}{Integer which dictates which species in the occupancy vector you want to test likelihood of.}

\item{ssv}{Whether or not the model contains species specific variables.}
}
\value{
Value with likelihood of species at that site.
}
\description{
Calculates conditional likelihood of one species presences based on environmental values and presence of other species.
}
