% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_cond_entire.R
\name{pred_cond_entire}
\alias{pred_cond_entire}
\title{Conditional prediction of species at range of sites site.}
\usage{
pred_cond_entire(draws, betas_shared, betas_ssv = NULL, site_matrix,
  occupancy_matrix, temp_vector, focal_species, ssv = FALSE)
}
\arguments{
\item{draws}{A set of mcmc draws produced by the GRETA function draws.}

\item{betas_shared}{Non species-specific enrivonmental parameters.}

\item{betas_ssv}{Species-specific environmental parameters.}

\item{occupancy_matrix}{Occupancy matrix for all species.}

\item{temp_vector}{Vector of temperatures at different sites, used to determine cooccurrence matrices.}

\item{focal_species}{Integer which dictates which species in the occupancy vector you want to test likelihood of.}

\item{ssv}{Whether or not the model contains species specific variables.}
}
\value{
Vector of length site giving likelihoods of species presence at every site.
}
\description{
Calculates conditional likelihood of one species presences based on environmental values and presence of other species at a range of sites.
}
