% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_correlation.R
\name{pred_correlation}
\alias{pred_correlation}
\title{Predictive species association matrix}
\usage{
pred_correlation(draws, X_assoc_pred, species_names)
}
\arguments{
\item{X_assoc_pred}{Vector of previously defined values, which represent standard deviations to or from the mean of the environmental parameter you are monitoring changes over.}

\item{species_names}{Names of species in model.}

\item{darws}{Series of mcmc draws produced by the GRETA function draws.}
}
\value{
A species x species matrix with assocation means for species associations when an environmental parameter is of a certain value.
}
\description{
Creates sppecies cooccurrence matrix at certain value of environmental parameter.
}
