% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_species_perc.R
\name{predict_species_perc}
\alias{predict_species_perc}
\title{Marginal prediction of species likelihoods at sites.}
\usage{
predict_species_perc(draws, X_newdata, site_id_newdata = NULL,
  X_assoc_newdata = NULL, species_list)
}
\arguments{
\item{draws}{A set of mcmc draws produced by the GRETA function draws.}

\item{X_newdata}{Data for new sites to predict for.}

\item{site_id_newdata}{Catchment IDs for new sites.}

\item{X_assoc_newdata}{Avergae association correlation matrix.}

\item{species_list}{List of species names.}
}
\value{
Matrix of site x species with likelihoods of presence at every site.
}
\description{
Calculates likelihood of species presences based solely on environmental values (marignal prediction).
}
